Separate request utils from database mapping utils
